changes on bo localisation controller

+++ WorkiSiteWeb/Worki.Database/dbo/Tables/MemberBooking.sql
@@ -5,7 +5,7 @@
     [FromDate]       DATETIME       NOT NULL,
     [ToDate]         DATETIME       NOT NULL, 
     [Message]        NVARCHAR (MAX) NULL,
-    [Price]          INT            DEFAULT (0) NOT NULL,
+    [Price]          DECIMAL(18, 2)            DEFAULT (0) NOT NULL,
     [StatusId]	     INT NOT NULL   DEFAULT (0),
     [Response]       NVARCHAR(MAX)  NULL, 
     CONSTRAINT [FK_MemberBooking_Offer] FOREIGN KEY ([OfferId]) REFERENCES [dbo].[Offer] ([Id]) ON DELETE CASCADE ON UPDATE NO ACTION,
